+++ docs/changelog.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-03-16
+
+- Add security documentation.
+
+## 2025-03-12
+
+- Updated charm code to trigger certificate refresh conditionally rather than through refresh hooks.
+
 ## 2025-02-26
 
 - Add some more logs and unit tests on the `proxied_endpoints` method.